Move helper function to avoid duplicates across scripts
